refactor(quartzite-core): move connect tests to integration file to fix file-size limit
quartzite-core/src/connect.rs grew to 1776 lines total (hard limit 1500)
after adding arity-relaxation fixtures and tests. Extract all tests that do
not need private crate::signal::tests access to quartzite-core/tests/connect.rs.
The inline mod tests now holds only Sender/Receiver fixtures and the
auto_cross_thread_posts_to_dispatcher test that requires private dispatcher
helpers. Result: src/connect.rs 720 lines; tests/connect.rs 1201 lines.
