+++ ai-docs/key-decisions.md
@@ -96,7 +96,7 @@ Driver callback reads `TimerState::signals_blocked: AtomicBool` (no `&mut Timer`
 
 ## Signal-to-signal connections
 
-Two APIs in `quartzite-core::connect` (std-only, in `quartzite::prelude`): `connect_signal_to_signal(&mut dyn Object, …, Arc<Mutex<dyn Object>>, …)` (dynamic, any `Object`) and `connect_signals<From, To, Args>(…, fn(&mut From) → &mut Signal<Args>, …)` (typed, avoids `&[Value]` round-trip for Auto/Queued). Validation at connection time via `SignalMeta::params` arity + `type_name` string comparison. `Object::emit_signal(&mut self, name, &[Value])` method added to the `Object` trait — proc-macro generates a hidden helper `__emit_signal_T` that dispatches by name and converts `Value` back to typed args via `FromValue`. Liveness via `Weak<Mutex<dyn Object>>` inside the forwarding closure.
+Two APIs in `quartzite-core::connect` (std-only, in `quartzite::prelude`): `connect_signal_to_signal(&mut dyn Object, …, Arc<Mutex<dyn Object>>, …)` (dynamic, any `Object`) and `connect_signals<From, To, Args>(…, fn(&mut From) → &mut Signal<Args>, …)` (typed, avoids `&[Value]` round-trip for Auto/Queued). Validation at connection time: source arity must be ≥ target arity (`from_arity < to_arity` → `ArityMismatch`); `type_name` strings are compared on the retained prefix; extra trailing source arguments are dropped at emit time (`&args[..to_arity]`). `connect_signal_to_slot` additionally validates against `MethodMeta::params` when the slot is meta-registered (no-meta fallback preserves `&[]` invocation). `Object::emit_signal(&mut self, name, &[Value])` method added to the `Object` trait — proc-macro generates a hidden helper `__emit_signal_T` that dispatches by name and converts `Value` back to typed args via `FromValue`. Liveness via `Weak<Mutex<dyn Object>>` inside the forwarding closure.
 
 ## `Timer` implements `Object`
