+++ quartzite-core/src/connect.rs
@@ -496,16 +496,13 @@ where
     Ok(id)
 }
 
+/// Tests that require access to `crate::signal::tests` (a `#[cfg(test)]`-private
+/// dispatcher helper) live here. All other connect tests are in
+/// `quartzite-core/tests/connect.rs` (integration tests) to keep this file
+/// within the 1000-line production / 1500-line total hard limit.
 #[cfg(test)]
 mod tests {
     use super::*;
-    use std::sync::{
-        Arc,
-        atomic::{AtomicI32, Ordering},
-    };
-
-    use parking_lot::Mutex;
-
     use crate::{
         id::ConnectionId,
         meta::MetaObject,
@@ -513,13 +510,16 @@ mod tests {
         signal::{ConnectionType, Signal},
         value::Value,
     };
+    use parking_lot::Mutex;
+    use std::sync::{
+        Arc,
+        atomic::{AtomicI32, Ordering},
+    };
 
-    // Minimal hand-written object with two signals for testing.
     struct Sender {
         base: ObjectBase,
         pub sig_a: Signal<(i32,)>,
     }
-
     impl Sender {
         fn new() -> Self {
             Self {
@@ -528,7 +528,6 @@ mod tests {
             }
         }
     }
-
     impl crate::AsObject for Sender {
         fn object_base(&self) -> &ObjectBase {
             &self.base
@@ -543,7 +542,6 @@ mod tests {
             self
         }
     }
-
     static SENDER_PARAMS: [crate::meta::ParamMeta; 1] =
         [crate::meta::ParamMeta::new("arg0", "i32")];
     static SENDER_SIGNALS: [crate::meta::SignalMeta; 1] =
@@ -565,7 +563,6 @@ mod tests {
         crate::meta::noop_lookup_method,
         crate::meta::noop_lookup_enum,
     );
-
     impl crate::Object for Sender {
         fn meta_object(&self) -> &'static MetaObject {
             &SENDER_META
@@ -704,1016 +701,6 @@ mod tests {
                 _ => None,
             }
         }
-    }
-
-    #[test]
-    fn unknown_from_signal_returns_error() {
-        let mut sender = Sender::new();
-        let to = Arc::new(Mutex::new(Receiver::new())) as Arc<Mutex<dyn Object>>;
-        let err = connect_signal_to_signal(
-            &mut sender,
-            "nonexistent",
-            &to,
-            "sig_b",
-            ConnectionType::Direct,
-        )
-        .unwrap_err();
-        assert_eq!(
-            err,
-            SignalConnectionError::UnknownFromSignal("nonexistent".into())
-        );
-    }
-
-    #[test]
-    fn unknown_to_signal_returns_error() {
-        let mut sender = Sender::new();
-        let to = Arc::new(Mutex::new(Receiver::new())) as Arc<Mutex<dyn Object>>;
-        let err = connect_signal_to_signal(
-            &mut sender,
-            "sig_a",
-            &to,
-            "nonexistent",
-            ConnectionType::Direct,
-        )
-        .unwrap_err();
-        assert_eq!(
-            err,
-            SignalConnectionError::UnknownToSignal("nonexistent".into())
-        );
-    }
-
-    // Zero-arg signal object used by arity tests as a source or recording target.
-    struct NullRecv {
-        base: ObjectBase,
-        pub sig: Signal<()>,
-    }
-    static NULL_SIGS: [crate::meta::SignalMeta; 1] = [crate::meta::SignalMeta::new("sig", &[])];
-    static NULL_META: MetaObject = MetaObject::new(
-        "NullRecv",
-        &[],
-        &NULL_SIGS,
-        &[],
-        &[],
-        crate::meta::noop_lookup_property,
-        |name| {
-            if name == "sig" {
-                Some(NULL_SIGS[0])
-            } else {
-                None
-            }
-        },
-        crate::meta::noop_lookup_method,
-        crate::meta::noop_lookup_enum,
-    );
-    impl crate::AsObject for NullRecv {
-        fn object_base(&self) -> &ObjectBase {
-            &self.base
-        }
-        fn object_base_mut(&mut self) -> &mut ObjectBase {
-            &mut self.base
-        }
-        fn as_any(&self) -> &dyn core::any::Any {
-            self
-        }
-        fn as_any_mut(&mut self) -> &mut dyn core::any::Any {
-            self
-        }
-    }
-    impl crate::Object for NullRecv {
-        fn meta_object(&self) -> &'static MetaObject {
-            &NULL_META
-        }
-        fn read_property(&self, _: &str) -> Option<Value> {
-            None
-        }
-        fn write_property(&mut self, _: &str, _: Value) -> bool {
-            false
-        }
-        fn invoke_method(&mut self, _: &str, _: &[Value]) -> Option<Value> {
-            None
-        }
-        fn connect_signal(
-            &mut self,
-            signal: &str,
-            callback: crate::traits::SignalCallback,
-            conn_type: ConnectionType,
-        ) -> Option<ConnectionId> {
-            match signal {
-                "sig" => {
-                    let cb = Arc::new(callback);
-                    Some(
-                        self.sig
-                            .connect_typed(move |_args: &()| (*cb)(&[]), conn_type),
-                    )
-                }
-                _ => None,
-            }
-        }
-        fn emit_signal(&mut self, signal: &str, args: &[Value]) -> Option<()> {
-            match signal {
-                "sig" => {
-                    if args.is_empty() {
-                        crate::emit!(self.sig, &());
-                        Some(())
-                    } else {
-                        None
-                    }
-                }
-                _ => None,
-            }
-        }
-    }
-
-    // Recording variant of NullRecv: sets an AtomicBool flag when `emit_signal("sig", _)` is
-    // called. Used to verify the forwarding callback fires after truncation.
-    struct RecordingNullRecv {
-        base: ObjectBase,
-        pub sig: Signal<()>,
-        pub fired: Arc<std::sync::atomic::AtomicBool>,
-    }
-    impl crate::AsObject for RecordingNullRecv {
-        fn object_base(&self) -> &ObjectBase {
-            &self.base
-        }
-        fn object_base_mut(&mut self) -> &mut ObjectBase {
-            &mut self.base
-        }
-        fn as_any(&self) -> &dyn core::any::Any {
-            self
-        }
-        fn as_any_mut(&mut self) -> &mut dyn core::any::Any {
-            self
-        }
-    }
-    impl crate::Object for RecordingNullRecv {
-        fn meta_object(&self) -> &'static MetaObject {
-            // Shares NullRecv's meta: one 0-arg signal named "sig".
-            &NULL_META
-        }
-        fn read_property(&self, _: &str) -> Option<Value> {
-            None
-        }
-        fn write_property(&mut self, _: &str, _: Value) -> bool {
-            false
-        }
-        fn invoke_method(&mut self, _: &str, _: &[Value]) -> Option<Value> {
-            None
-        }
-        fn connect_signal(
-            &mut self,
-            signal: &str,
-            callback: crate::traits::SignalCallback,
-            conn_type: ConnectionType,
-        ) -> Option<ConnectionId> {
-            match signal {
-                "sig" => {
-                    let cb = Arc::new(callback);
-                    Some(
-                        self.sig
-                            .connect_typed(move |_args: &()| (*cb)(&[]), conn_type),
-                    )
-                }
-                _ => None,
-            }
-        }
-        fn emit_signal(&mut self, signal: &str, args: &[Value]) -> Option<()> {
-            match signal {
-                "sig" => {
-                    if args.is_empty() {
-                        self.fired.store(true, std::sync::atomic::Ordering::Relaxed);
-                        crate::emit!(self.sig, &());
-                        Some(())
-                    } else {
-                        None
-                    }
-                }
-                _ => None,
-            }
-        }
-    }
-
-    #[test]
-    fn arity_mismatch_returns_error() {
-        // New rule: from_arity < to_arity is the only rejection.
-        // Use a 0-arg source (NullRecv) → 1-arg target (Receiver) to keep the rejection path
-        // exercised: ArityMismatch { from: 0, to: 1 }.
-        let mut source = NullRecv {
-            base: ObjectBase::new(),
-            sig: Signal::new(),
-        };
-        let to = Arc::new(Mutex::new(Receiver::new())) as Arc<Mutex<dyn Object>>;
-        let err =
-            connect_signal_to_signal(&mut source, "sig", &to, "sig_b", ConnectionType::Direct)
-                .unwrap_err();
-        assert_eq!(err, SignalConnectionError::ArityMismatch { from: 0, to: 1 });
-    }
-
-    #[test]
-    fn arity_relaxation_truncates_extras() {
-        // 1-arg source (Sender / sig_a) → 0-arg target (RecordingNullRecv / sig).
-        // from_arity (1) >= to_arity (0): connect succeeds and the target fires on emit.
-        let fired = Arc::new(std::sync::atomic::AtomicBool::new(false));
-        let mut sender = Sender::new();
-        let rnr = RecordingNullRecv {
-            base: ObjectBase::new(),
-            sig: Signal::new(),
-            fired: Arc::clone(&fired),
-        };
-        let to = Arc::new(Mutex::new(rnr)) as Arc<Mutex<dyn Object>>;
-        connect_signal_to_signal(&mut sender, "sig_a", &to, "sig", ConnectionType::Direct)
-            .expect("connection must succeed: from_arity (1) >= to_arity (0)");
-        sender.sig_a.emit_unconditionally(&(42,));
-        assert!(
-            fired.load(std::sync::atomic::Ordering::Relaxed),
-            "target emit_signal must have been called after truncation"
-        );
-    }
-
-    #[test]
-    fn type_mismatch_returns_error() {
-        struct BoolRecv {
-            base: ObjectBase,
-            pub _sig: Signal<(bool,)>,
-        }
-        static BOOL_PARAMS: [crate::meta::ParamMeta; 1] =
-            [crate::meta::ParamMeta::new("arg0", "bool")];
-        static BOOL_SIGS: [crate::meta::SignalMeta; 1] =
-            [crate::meta::SignalMeta::new("sig", &BOOL_PARAMS)];
-        static BOOL_META: MetaObject = MetaObject::new(
-            "BoolRecv",
-            &[],
-            &BOOL_SIGS,
-            &[],
-            &[],
-            crate::meta::noop_lookup_property,
-            |name| {
-                if name == "sig" {
-                    Some(BOOL_SIGS[0])
-                } else {
-                    None
-                }
-            },
-            crate::meta::noop_lookup_method,
-            crate::meta::noop_lookup_enum,
-        );
-        impl crate::AsObject for BoolRecv {
-            fn object_base(&self) -> &ObjectBase {
-                &self.base
-            }
-            fn object_base_mut(&mut self) -> &mut ObjectBase {
-                &mut self.base
-            }
-            fn as_any(&self) -> &dyn core::any::Any {
-                self
-            }
-            fn as_any_mut(&mut self) -> &mut dyn core::any::Any {
-                self
-            }
-        }
-        impl crate::Object for BoolRecv {
-            fn meta_object(&self) -> &'static MetaObject {
-                &BOOL_META
-            }
-            fn read_property(&self, _: &str) -> Option<Value> {
-                None
-            }
-            fn write_property(&mut self, _: &str, _: Value) -> bool {
-                false
-            }
-            fn invoke_method(&mut self, _: &str, _: &[Value]) -> Option<Value> {
-                None
-            }
-            fn connect_signal(
-                &mut self,
-                _signal: &str,
-                _callback: crate::traits::SignalCallback,
-                _conn_type: ConnectionType,
-            ) -> Option<ConnectionId> {
-                None
-            }
-            fn emit_signal(&mut self, _: &str, _: &[Value]) -> Option<()> {
-                None
-            }
-        }
-        let mut sender = Sender::new();
-        let br = BoolRecv {
-            base: ObjectBase::new(),
-            _sig: Signal::new(),
-        };
-        let to = Arc::new(Mutex::new(br)) as Arc<Mutex<dyn Object>>;
-        let err =
-            connect_signal_to_signal(&mut sender, "sig_a", &to, "sig", ConnectionType::Direct)
-                .unwrap_err();
-        assert_eq!(
-            err,
-            SignalConnectionError::TypeMismatch {
-                index: 0,
-                from: "i32".into(),
-                to: "bool".into()
-            }
-        );
-    }
-
-    #[test]
-    fn direct_connection_forwards_signal() {
-        let mut sender = Sender::new();
-        let receiver = Arc::new(Mutex::new(Receiver::new()));
-        let counter = Arc::new(AtomicI32::new(0));
-        {
-            let c = Arc::clone(&counter);
-            receiver.lock().sig_b.connect(move |args: &(i32,)| {
-                c.store(args.0, Ordering::Relaxed);
-            });
-        }
-        let to: Arc<Mutex<dyn Object>> = Arc::clone(&receiver) as Arc<Mutex<dyn Object>>;
-        let id =
-            connect_signal_to_signal(&mut sender, "sig_a", &to, "sig_b", ConnectionType::Direct)
-                .expect("connection must succeed");
-
-        sender.sig_a.emit_unconditionally(&(42,));
-        assert_eq!(
-            counter.load(Ordering::Relaxed),
-            42,
-            "value must be forwarded"
-        );
-
-        // AC8: disconnect stops forwarding.
-        sender.sig_a.disconnect(id);
-        sender.sig_a.emit_unconditionally(&(99,));
-        assert_eq!(
-            counter.load(Ordering::Relaxed),
-            42,
-            "value must not change after disconnect"
-        );
-    }
-
-    #[test]
-    fn single_shot_fires_once_and_slot_is_removed() {
-        let mut sender = Sender::new();
-        let receiver = Arc::new(Mutex::new(Receiver::new()));
-        let counter = Arc::new(AtomicI32::new(0));
-        {
-            let c = Arc::clone(&counter);
-            receiver
-                .lock()
-                .sig_b
-                .connect(move |args: &(i32,)| c.store(args.0, Ordering::Relaxed));
-        }
-        let to: Arc<Mutex<dyn Object>> = Arc::clone(&receiver) as Arc<Mutex<dyn Object>>;
-        connect_signal_to_signal(
-            &mut sender,
-            "sig_a",
-            &to,
-            "sig_b",
-            ConnectionType::SingleShot,
-        )
-        .expect("connection must succeed");
-
-        sender.sig_a.emit_unconditionally(&(1,));
-        assert_eq!(
-            counter.load(Ordering::Relaxed),
-            1,
-            "must fire on first emit"
-        );
-
-        sender.sig_a.emit_unconditionally(&(2,));
-        assert_eq!(
-            counter.load(Ordering::Relaxed),
-            1,
-            "slot must not fire on second emit — auto-disconnected after first delivery"
-        );
-    }
-
-    #[test]
-    fn liveness_silently_drops_when_to_arc_released() {
-        let mut sender = Sender::new();
-        let receiver = Arc::new(Mutex::new(Receiver::new()));
-        let counter = Arc::new(AtomicI32::new(0));
-        {
-            let c = Arc::clone(&counter);
-            receiver.lock().sig_b.connect(move |args: &(i32,)| {
-                c.store(args.0, Ordering::Relaxed);
-            });
-        }
-        let to: Arc<Mutex<dyn Object>> = Arc::clone(&receiver) as Arc<Mutex<dyn Object>>;
-        connect_signal_to_signal(&mut sender, "sig_a", &to, "sig_b", ConnectionType::Direct)
-            .unwrap();
-        // Verify it fires before drop.
-        sender.sig_a.emit_unconditionally(&(1,));
-        assert_eq!(counter.load(Ordering::Relaxed), 1);
-        // Drop all strong Arcs — Weak inside the callback now returns None.
-        drop(to);
-        drop(receiver);
-        // AC7: no panic, connection silently skipped.
-        sender.sig_a.emit_unconditionally(&(2,));
-        assert_eq!(
-            counter.load(Ordering::Relaxed),
-            1,
-            "no forwarding after to is dropped"
-        );
-    }
-
-    #[test]
-    fn connect_signals_typed_direct_forwards() {
-        let mut sender = Sender::new();
-        let receiver = Arc::new(Mutex::new(Receiver::new()));
-        let counter = Arc::new(AtomicI32::new(0));
-        {
-            let c = Arc::clone(&counter);
-            receiver.lock().sig_b.connect(move |args: &(i32,)| {
-                c.store(args.0, Ordering::Relaxed);
-            });
-        }
-        let id = connect_signals(
-            &mut sender,
-            "sig_a",
-            |obj: &mut Sender| &mut obj.sig_a,
-            &receiver,
-            "sig_b",
-            ConnectionType::Direct,
-        )
-        .expect("typed connection must succeed");
-
-        sender.sig_a.emit_unconditionally(&(7,));
-        assert_eq!(counter.load(Ordering::Relaxed), 7);
-        sender.sig_a.disconnect(id);
-        sender.sig_a.emit_unconditionally(&(8,));
-        assert_eq!(
-            counter.load(Ordering::Relaxed),
-            7,
-            "no forwarding after disconnect"
-        );
-    }
-
-    #[test]
-    fn connect_signals_typed_liveness() {
-        let mut sender = Sender::new();
-        let receiver = Arc::new(Mutex::new(Receiver::new()));
-        let counter = Arc::new(AtomicI32::new(0));
-        {
-            let c = Arc::clone(&counter);
-            receiver.lock().sig_b.connect(move |args: &(i32,)| {
-                c.store(args.0, Ordering::Relaxed);
-            });
-        }
-        connect_signals(
-            &mut sender,
-            "sig_a",
-            |obj: &mut Sender| &mut obj.sig_a,
-            &receiver,
-            "sig_b",
-            ConnectionType::Direct,
-        )
-        .unwrap();
-        sender.sig_a.emit_unconditionally(&(3,));
-        assert_eq!(counter.load(Ordering::Relaxed), 3);
-        drop(receiver);
-        sender.sig_a.emit_unconditionally(&(4,));
-        assert_eq!(
-            counter.load(Ordering::Relaxed),
-            3,
-            "no forwarding after drop"
-        );
-    }
-
-    // AC6 — Auto same-thread: fires synchronously (no dispatcher needed).
-    #[test]
-    fn auto_same_thread_fires_synchronously() {
-        let mut sender = Sender::new();
-        let receiver = Arc::new(Mutex::new(Receiver::new()));
-        let counter = Arc::new(AtomicI32::new(0));
-        {
-            let c = Arc::clone(&counter);
-            receiver
-                .lock()
-                .sig_b
-                .connect(move |args: &(i32,)| c.store(args.0, Ordering::Relaxed));
-        }
-        let to: Arc<Mutex<dyn Object>> = Arc::clone(&receiver) as Arc<Mutex<dyn Object>>;
-        connect_signal_to_signal(&mut sender, "sig_a", &to, "sig_b", ConnectionType::Auto).unwrap();
-
-        sender.sig_a.emit_unconditionally(&(55,));
-        assert_eq!(
-            counter.load(Ordering::Relaxed),
-            55,
-            "AC6 same-thread Auto must fire synchronously"
-        );
-    }
-
-    // Two-arg sender used by typed-path arity and truncation tests.
-    struct Sender2 {
-        base: ObjectBase,
-        pub sig: Signal<(i32, i32)>,
-    }
-    impl Sender2 {
-        fn new() -> Self {
-            Self {
-                base: ObjectBase::new(),
-                sig: Signal::new(),
-            }
-        }
-    }
-    static S2_PARAMS: [crate::meta::ParamMeta; 2] = [
-        crate::meta::ParamMeta::new("a", "i32"),
-        crate::meta::ParamMeta::new("b", "i32"),
-    ];
-    static S2_SIGS: [crate::meta::SignalMeta; 1] =
-        [crate::meta::SignalMeta::new("sig", &S2_PARAMS)];
-    static S2_META: MetaObject = MetaObject::new(
-        "Sender2",
-        &[],
-        &S2_SIGS,
-        &[],
-        &[],
-        crate::meta::noop_lookup_property,
-        |name| {
-            if name == "sig" {
-                Some(S2_SIGS[0])
-            } else {
-                None
-            }
-        },
-        crate::meta::noop_lookup_method,
-        crate::meta::noop_lookup_enum,
-    );
-    impl crate::AsObject for Sender2 {
-        fn object_base(&self) -> &ObjectBase {
-            &self.base
-        }
-        fn object_base_mut(&mut self) -> &mut ObjectBase {
-            &mut self.base
-        }
-        fn as_any(&self) -> &dyn core::any::Any {
-            self
-        }
-        fn as_any_mut(&mut self) -> &mut dyn core::any::Any {
-            self
-        }
-    }
-    impl crate::Object for Sender2 {
-        fn meta_object(&self) -> &'static MetaObject {
-            &S2_META
-        }
-        fn read_property(&self, _: &str) -> Option<Value> {
-            None
-        }
-        fn write_property(&mut self, _: &str, _: Value) -> bool {
-            false
-        }
-        fn invoke_method(&mut self, _: &str, _: &[Value]) -> Option<Value> {
-            None
-        }
-        fn connect_signal(
-            &mut self,
-            signal: &str,
-            callback: crate::traits::SignalCallback,
-            conn_type: ConnectionType,
-        ) -> Option<ConnectionId> {
-            match signal {
-                "sig" => {
-                    let cb = Arc::new(callback);
-                    Some(self.sig.connect_typed(
-                        move |args: &(i32, i32)| {
-                            (*cb)(&[
-                                crate::IntoValue::into_value(args.0),
-                                crate::IntoValue::into_value(args.1),
-                            ]);
-                        },
-                        conn_type,
-                    ))
-                }
-                _ => None,
-            }
-        }
-        fn emit_signal(&mut self, signal: &str, args: &[Value]) -> Option<()> {
-            match signal {
-                "sig" => {
-                    if args.len() != 2 {
-                        return None;
-                    }
-                    let a: i32 = crate::FromValue::from_value(args[0].clone()).ok()?;
-                    let b: i32 = crate::FromValue::from_value(args[1].clone()).ok()?;
-                    crate::emit!(self.sig, &(a, b));
-                    Some(())
-                }
-                _ => None,
-            }
-        }
-    }
-
-    // Two-arg receiver used by the flipped connect_signals_typed arity-mismatch test.
-    struct BigReceiver {
-        base: ObjectBase,
-        pub _sig: Signal<(i32, i32)>,
-    }
-    static BIG_RECV_PARAMS: [crate::meta::ParamMeta; 2] = [
-        crate::meta::ParamMeta::new("a", "i32"),
-        crate::meta::ParamMeta::new("b", "i32"),
-    ];
-    static BIG_RECV_SIGS: [crate::meta::SignalMeta; 1] =
-        [crate::meta::SignalMeta::new("sig", &BIG_RECV_PARAMS)];
-    static BIG_RECV_META: MetaObject = MetaObject::new(
-        "BigReceiver",
-        &[],
-        &BIG_RECV_SIGS,
-        &[],
-        &[],
-        crate::meta::noop_lookup_property,
-        |name| {
-            if name == "sig" {
-                Some(BIG_RECV_SIGS[0])
-            } else {
-                None
-            }
-        },
-        crate::meta::noop_lookup_method,
-        crate::meta::noop_lookup_enum,
-    );
-    impl crate::AsObject for BigReceiver {
-        fn object_base(&self) -> &ObjectBase {
-            &self.base
-        }
-        fn object_base_mut(&mut self) -> &mut ObjectBase {
-            &mut self.base
-        }
-        fn as_any(&self) -> &dyn core::any::Any {
-            self
-        }
-        fn as_any_mut(&mut self) -> &mut dyn core::any::Any {
-            self
-        }
-    }
-    impl crate::Object for BigReceiver {
-        fn meta_object(&self) -> &'static MetaObject {
-            &BIG_RECV_META
-        }
-        fn read_property(&self, _: &str) -> Option<Value> {
-            None
-        }
-        fn write_property(&mut self, _: &str, _: Value) -> bool {
-            false
-        }
-        fn invoke_method(&mut self, _: &str, _: &[Value]) -> Option<Value> {
-            None
-        }
-        fn connect_signal(
-            &mut self,
-            _s: &str,
-            _cb: crate::traits::SignalCallback,
-            _ct: ConnectionType,
-        ) -> Option<ConnectionId> {
-            None
-        }
-        fn emit_signal(&mut self, _: &str, _: &[Value]) -> Option<()> {
-            None
-        }
-    }
-
-    // R2 partial — connect_signals arity / type mismatch error arms
-    // (connect_signal_to_signal already has arity/type tests; these cover
-    // the same checks inside the typed connect_signals wrapper.)
-
-    #[test]
-    fn connect_signals_typed_arity_mismatch_returns_error() {
-        // New rule: from_arity < to_arity is the only rejection.
-        // Use Sender (1-arg sig_a) → BigReceiver (2-arg sig) to keep rejection exercised:
-        // ArityMismatch { from: 1, to: 2 }.
-        let mut sender = Sender::new();
-        let big_recv = Arc::new(Mutex::new(BigReceiver {
-            base: ObjectBase::new(),
-            _sig: Signal::new(),
-        }));
-        let err = connect_signals::<_, _, (i32,)>(
-            &mut sender,
-            "sig_a",
-            |obj: &mut Sender| &mut obj.sig_a,
-            &big_recv,
-            "sig",
-            ConnectionType::Direct,
-        )
-        .unwrap_err();
-        assert_eq!(err, SignalConnectionError::ArityMismatch { from: 1, to: 2 });
-    }
-
-    #[test]
-    fn connect_signals_typed_truncates_extras() {
-        // 2-arg source (Sender2 / sig) → 1-arg target (Receiver / sig_b).
-        // from_arity (2) >= to_arity (1): connect succeeds; only the first arg is forwarded.
-        let mut s2 = Sender2::new();
-        let receiver = Arc::new(Mutex::new(Receiver::new()));
-        let captured = Arc::new(AtomicI32::new(-1));
-        {
-            let c = Arc::clone(&captured);
-            receiver
-                .lock()
-                .sig_b
-                .connect(move |args: &(i32,)| c.store(args.0, Ordering::Relaxed));
-        }
-        connect_signals::<_, _, (i32, i32)>(
-            &mut s2,
-            "sig",
-            |obj: &mut Sender2| &mut obj.sig,
-            &receiver,
-            "sig_b",
-            ConnectionType::Direct,
-        )
-        .expect("connection must succeed: from_arity (2) >= to_arity (1)");
-
-        s2.sig.emit_unconditionally(&(11, 99));
-        assert_eq!(
-            captured.load(Ordering::Relaxed),
-            11,
-            "only the first arg must be forwarded; the second is dropped"
-        );
-    }
-
-    #[test]
-    fn connect_signals_typed_type_mismatch_returns_error() {
-        // Sender has sig_a: Signal<(i32,)> — type "i32".
-        // Build a receiver with sig carrying type "bool" to trigger TypeMismatch.
-        struct BoolRecv2 {
-            base: ObjectBase,
-            _sig: Signal<(bool,)>,
-        }
-        static BR2_PARAMS: [crate::meta::ParamMeta; 1] =
-            [crate::meta::ParamMeta::new("arg0", "bool")];
-        static BR2_SIGS: [crate::meta::SignalMeta; 1] =
-            [crate::meta::SignalMeta::new("sig", &BR2_PARAMS)];
-        static BR2_META: MetaObject = MetaObject::new(
-            "BoolRecv2",
-            &[],
-            &BR2_SIGS,
-            &[],
-            &[],
-            crate::meta::noop_lookup_property,
-            |name| {
-                if name == "sig" {
-                    Some(BR2_SIGS[0])
-                } else {
-                    None
-                }
-            },
-            crate::meta::noop_lookup_method,
-            crate::meta::noop_lookup_enum,
-        );
-        impl crate::AsObject for BoolRecv2 {
-            fn object_base(&self) -> &ObjectBase {
-                &self.base
-            }
-            fn object_base_mut(&mut self) -> &mut ObjectBase {
-                &mut self.base
-            }
-            fn as_any(&self) -> &dyn core::any::Any {
-                self
-            }
-            fn as_any_mut(&mut self) -> &mut dyn core::any::Any {
-                self
-            }
-        }
-        impl crate::Object for BoolRecv2 {
-            fn meta_object(&self) -> &'static MetaObject {
-                &BR2_META
-            }
-            fn read_property(&self, _: &str) -> Option<Value> {
-                None
-            }
-            fn write_property(&mut self, _: &str, _: Value) -> bool {
-                false
-            }
-            fn invoke_method(&mut self, _: &str, _: &[Value]) -> Option<Value> {
-                None
-            }
-            fn connect_signal(
-                &mut self,
-                _s: &str,
-                _cb: crate::traits::SignalCallback,
-                _ct: ConnectionType,
-            ) -> Option<ConnectionId> {
-                None
-            }
-            fn emit_signal(&mut self, _: &str, _: &[Value]) -> Option<()> {
-                None
-            }
-        }
-        let mut sender = Sender::new();
-        let br2 = Arc::new(Mutex::new(BoolRecv2 {
-            base: ObjectBase::new(),
-            _sig: Signal::new(),
-        }));
-        let err = connect_signals::<_, _, (i32,)>(
-            &mut sender,
-            "sig_a",
-            |obj: &mut Sender| &mut obj.sig_a,
-            &br2,
-            "sig",
-            ConnectionType::Direct,
-        )
-        .unwrap_err();
-        assert_eq!(
-            err,
-            SignalConnectionError::TypeMismatch {
-                index: 0,
-                from: "i32".into(),
-                to: "bool".into()
-            }
-        );
-    }
-
-    // Hand-rolled Object whose invoke_method records the received args slice.
-    // Used by connect_signal_to_slot meta-validated-path tests (subtask 6).
-    struct RecordingSlotRecv {
-        base: ObjectBase,
-        meta: &'static MetaObject,
-        pub recorded: Arc<Mutex<Vec<Value>>>,
-    }
-    impl RecordingSlotRecv {
-        fn new(meta: &'static MetaObject, recorded: Arc<Mutex<Vec<Value>>>) -> Self {
-            Self {
-                base: ObjectBase::new(),
-                meta,
-                recorded,
-            }
-        }
-    }
-    // Three static MetaObject instances with differing lookup_method behaviour.
-    // (a) Validated 1-arg i32 slot — lookup_method returns Some for "on_click".
-    static RSR_PARAMS_1ARG: [crate::meta::ParamMeta; 1] = [crate::meta::ParamMeta::new("v", "i32")];
-    static RSR_METHODS_1ARG: [crate::meta::MethodMeta; 1] = [crate::meta::MethodMeta::new(
-        "on_click",
-        &RSR_PARAMS_1ARG,
-        "()",
-    )];
-    static RECV_META_VALIDATED_1ARG: MetaObject = MetaObject::new(
-        "RecordingSlotRecv",
-        &[],
-        &[],
-        &RSR_METHODS_1ARG,
-        &[],
-        crate::meta::noop_lookup_property,
-        crate::meta::noop_lookup_signal,
-        |name| {
-            if name == "on_click" {
-                Some(RSR_METHODS_1ARG[0])
-            } else {
-                None
-            }
-        },
-        crate::meta::noop_lookup_enum,
-    );
-    // (b) Validated 1-arg bool slot.
-    static RSR_PARAMS_BOOL: [crate::meta::ParamMeta; 1] =
-        [crate::meta::ParamMeta::new("v", "bool")];
-    static RSR_METHODS_BOOL: [crate::meta::MethodMeta; 1] = [crate::meta::MethodMeta::new(
-        "on_click",
-        &RSR_PARAMS_BOOL,
-        "()",
-    )];
-    static RECV_META_VALIDATED_BOOL: MetaObject = MetaObject::new(
-        "RecordingSlotRecv",
-        &[],
-        &[],
-        &RSR_METHODS_BOOL,
-        &[],
-        crate::meta::noop_lookup_property,
-        crate::meta::noop_lookup_signal,
-        |name| {
-            if name == "on_click" {
-                Some(RSR_METHODS_BOOL[0])
-            } else {
-                None
-            }
-        },
-        crate::meta::noop_lookup_enum,
-    );
-    // (c) Fallback — noop_lookup_method; slot not in meta-system.
-    static RECV_META_FALLBACK: MetaObject = MetaObject::new(
-        "RecordingSlotRecv",
-        &[],
-        &[],
-        &[],
-        &[],
-        crate::meta::noop_lookup_property,
-        crate::meta::noop_lookup_signal,
-        crate::meta::noop_lookup_method,
-        crate::meta::noop_lookup_enum,
-    );
-    impl crate::AsObject for RecordingSlotRecv {
-        fn object_base(&self) -> &ObjectBase {
-            &self.base
-        }
-        fn object_base_mut(&mut self) -> &mut ObjectBase {
-            &mut self.base
-        }
-        fn as_any(&self) -> &dyn core::any::Any {
-            self
-        }
-        fn as_any_mut(&mut self) -> &mut dyn core::any::Any {
-            self
-        }
-    }
-    impl crate::Object for RecordingSlotRecv {
-        fn meta_object(&self) -> &'static MetaObject {
-            self.meta
-        }
-        fn read_property(&self, _: &str) -> Option<Value> {
-            None
-        }
-        fn write_property(&mut self, _: &str, _: Value) -> bool {
-            false
-        }
-        fn invoke_method(&mut self, _method: &str, args: &[Value]) -> Option<Value> {
-            *self.recorded.lock() = args.to_vec();
-            None
-        }
-        fn connect_signal(
-            &mut self,
-            _s: &str,
-            _cb: crate::traits::SignalCallback,
-            _ct: ConnectionType,
-        ) -> Option<ConnectionId> {
-            None
-        }
-        fn emit_signal(&mut self, _: &str, _: &[Value]) -> Option<()> {
-            None
-        }
-    }
-
-    #[test]
-    fn slot_arity_validation_truncates_args() {
-        // 2-arg source (Sender2 / sig) → RECV_META_VALIDATED_1ARG slot "on_click".
-        // from_arity (2) >= slot_arity (1): connect succeeds; slot is invoked with first arg only.
-        let recorded = Arc::new(Mutex::new(Vec::<Value>::new()));
-        let target = Arc::new(Mutex::new(RecordingSlotRecv::new(
-            &RECV_META_VALIDATED_1ARG,
-            Arc::clone(&recorded),
-        ))) as Arc<Mutex<dyn Object>>;
-        let mut s2 = Sender2::new();
-        connect_signal_to_slot(&mut s2, "sig", &target, "on_click")
-            .expect("connection must succeed: from_arity (2) >= slot_arity (1)");
-        s2.sig.emit_unconditionally(&(7, 8));
-        assert_eq!(
-            *recorded.lock(),
-            vec![Value::Int(7)],
-            "slot must be invoked with first arg only"
-        );
-    }
-
-    #[test]
-    fn slot_arity_mismatch_returns_error() {
-        // 0-arg source (NullRecv / sig) → RECV_META_VALIDATED_1ARG slot "on_click".
-        // from_arity (0) < slot_arity (1): ArityMismatch { from: 0, to: 1 }.
-        let recorded = Arc::new(Mutex::new(Vec::<Value>::new()));
-        let target = Arc::new(Mutex::new(RecordingSlotRecv::new(
-            &RECV_META_VALIDATED_1ARG,
-            Arc::clone(&recorded),
-        ))) as Arc<Mutex<dyn Object>>;
-        let mut source = NullRecv {
-            base: ObjectBase::new(),
-            sig: Signal::new(),
-        };
-        let err = connect_signal_to_slot(&mut source, "sig", &target, "on_click").unwrap_err();
-        assert_eq!(err, SignalConnectionError::ArityMismatch { from: 0, to: 1 });
-    }
-
-    #[test]
-    fn slot_type_mismatch_returns_error() {
-        // 1-arg i32 source (Sender / sig_a) → RECV_META_VALIDATED_BOOL slot "on_click".
-        // type_name "i32" != "bool": TypeMismatch { index: 0, from: "i32", to: "bool" }.
-        let recorded = Arc::new(Mutex::new(Vec::<Value>::new()));
-        let target = Arc::new(Mutex::new(RecordingSlotRecv::new(
-            &RECV_META_VALIDATED_BOOL,
-            Arc::clone(&recorded),
-        ))) as Arc<Mutex<dyn Object>>;
-        let mut sender = Sender::new();
-        let err = connect_signal_to_slot(&mut sender, "sig_a", &target, "on_click").unwrap_err();
-        assert_eq!(
-            err,
-            SignalConnectionError::TypeMismatch {
-                index: 0,
-                from: "i32".into(),
-                to: "bool".into()
-            }
-        );
-    }
-
-    #[test]
-    fn slot_meta_absent_falls_back_to_empty_args() {
-        // 1-arg source (Sender / sig_a) → RECV_META_FALLBACK slot "on_click".
-        // Slot not in meta-system: connection succeeds, slot invoked with &[] at emit time.
-        let recorded = Arc::new(Mutex::new(Vec::<Value>::new()));
-        let target = Arc::new(Mutex::new(RecordingSlotRecv::new(
-            &RECV_META_FALLBACK,
-            Arc::clone(&recorded),
-        ))) as Arc<Mutex<dyn Object>>;
-        let mut sender = Sender::new();
-        connect_signal_to_slot(&mut sender, "sig_a", &target, "on_click")
-            .expect("connection must succeed: fallback path (slot not in meta)");
-        sender.sig_a.emit_unconditionally(&(42,));
-        assert_eq!(
-            *recorded.lock(),
-            vec![],
-            "fallback path must invoke slot with empty args"
-        );
     }
 
     // AC6 — Auto cross-thread: callback is posted to the queued dispatcher.
